Note disabled in page popup.
On pages where the browser does not run extensions like vimium,
replace the popup with a note explaining what and why.

Fixes #2850.

Currently, this is unsatisfactory because, on the Chrome new tab page,
Vimium does not become active until the user tabs or clicks into the
page (which is weird).