Show browser-blocking complaint in the popup for empty portForTab object
To test the difference:
* load an unblocked URL (e.g. https://www.example.com) in a tab
* open the popup, see that the exclusion rules show as expected
* navigate the tab to a blocked URL (e.g. chrome://extensions)
* open the popup again

Before this commit, the popup still shows the exclusion rules, because
there is still a portForTabs object associated with the tabId.

This commit adds a second check to see if the object is empty, and the
message shows as expected, because all of the ports have been closed and
cleared from the object.