Fix Firefox link hints.
Fixes #2958 (probably).

This appears to be correct for target "_blank" (or not) and modifiers
(or not).  That's four cases.

It will be *incorrect* if there is a click listener on a target "_blank"
link.  Some conditions are tested to try to prevent that from happening.

This only affects Firefox.  Chrome is unaffected.