Add CodeMirror as special case in InsertMode
CodeMirror (a code editor component that can be embedded in web pages)
implements a simple Vim emulation mode. It supports navigating the text
with Vim normal-mode operators, and also editing text with insert-mode
operators. In particular, CodeMirror supports pressing ESC to go from
Insert mode to Normal mode.

To Vimium (and the DOM), CodeMirror appears as an editable, focusable
element. This means that when you press ESC in a CodeMirror editor,
Vimium will blur CodeMirror's focus. This prevents CodeMirror users from
using Vim mode.

This change allows CodeMirror's Vim keymap and Vimium to co-exist on the
same page. Without this, Vimium will never let a CodeMirror's editor
return to Normal mode. If we're inside a CodeMirror editor, instead of
trapping the ESC, we pass it through to the editor.

See https://github.com/philc/vimium/issues/2734 for more context.