Use hash (too) for local marks.
First, I *very rarely* use local marks.  They just don't seem as useful
in a browser as they are in a text editor.

However, on a page like GMail, I do often want to jump back and forward
between my labels.  These are different locations hashes (anchors).

The idea here is to make local marks useful for this case.

If the scroll positions (X and Y) are both 0, and the hash (anchor) is
present in the mark and non-empty, then change the hash instead of
scrolling.

On first tests, this appears to work nicely for changing labels in
GMail.  I'm sure there are other uses.

This change shouldn't interfere with use cases where the user has
scrolled within the page.