Bump minimum Chrome version to 69, to support ES2018
Since the removal of coffeescript in #3459, we're now using some newer
features of javascript. I think the most recent feature we're currently using is
Array.flat, which was introduced in Chrome 69 and Firefox 62.

I checked the google analytics for our chrome store page, and less than
0.5% of users visiting that page have a Chrome version prior to 80, so
this increased version requirement should be acceptable.

We don't currently enforce a minimum version for Firefox; I'll do that
in a separate commit.