Abandon our frankenstein use of modules, and use no modules
Now we assign every exported item to the `global` object, which is `global`
in node.js, and `window` in the browser (this is handled by utils.js).

Previously we were only using modules in tests, but then assigning all
exported members to the global namespace when importing, so we weren't
really using modules. Maybe this was required in the early days of node?

We can't embrace a proper use of ES6 modules, because they don't work in
chrome extension content scripts. So if we're not going to use them
there, let's not bother using them anywhere. It would just make the code
confusing if some of our code files used modules, and other files didn't.