Vendor shoulda.js
This simplifies the experience for new developers. The only reason we're
using git submodules is to track this library. This dependency is tiny
and changes rarely, so it's fine to include it directly in our repo, as
one would for a dependency like underscore.js or jquery.js.

Another strategy is to use npm module for this. I'm going to update
shoulda.js soon and publish it as an npm module.