Do not initialize the PostFindMode superclass when not focused on a textbox
Before the CS -> JS migration, this if statement was checked right at
the start of the PostFindMode constructor. In JS, this gives an error
because you cannot return before calling the superclass constructor.
However, calling that constructor before the check has unwanted side
effects. Extracing the if statement to a wrapper function that calls the
actual constructor fixes this.