Add command to open current tag in an incognito window
It is often useful to open a page in incognito mode, e.g. if you want to
see what a page looks like when not logged in. Instead of having to yank
the URL of the current tab, open a new incognito window, and paste the
URL, this new command combines those actions into one.

The new `openTabInIncognitoWindow` command is bound to `yI` by default
(analogous to `yt` for `duplicateTab`), and listed under "Advanced commands".