Fixed a minor handler-id assignment bug
Before the fix, the anonymous name, generated for assigning to an
otherwise unnamed handler, had a different value in the name than the id
assigned to the handler.

This bugfix makes sure that the id used in the name of an anonymous
handler matches the id assigned to the handler.