Don't trigger "InstallTrigger is deprecated and will be removed in the future" warning
By checking for presence in window it doesn't trigger the warning, which
is a bit annoying since it's in my console all the time and makes it
harder to see "did the script I'm working on throw a warning/error?" at
a glance.

Furthermore, it also puts back the old check that was removed in #3773;
InstallTrigger will be removed eventually, and it will then break for
everyone. This conditional should probably be updated, but I don't know
what to (and any updated version may have the same issue).

Fixes #4033