Move cursor at end of text on focused one-line textareas
When we `simulateSelect` an input and the cursor is at the start of the
text, we move it to the end of text for all `isSelectable` elements
*except* `textarea`s to avoid bad UX for multiline inputs (e.g. the end
of input could happen to be out of the viewport). However some sites may
use `<textarea>` for what really is a single-line input for 99% times of
the usage (e.g. Google search bar which recently has become a
`<textarea>`) and the cursor won't go to the end of text in such cases.

See:
    https://github.com/philc/vimium/issues/4247
    https://github.com/philc/vimium/issues/4251

This commit changes `simulateSelect` to skip `textarea` cursor
reposition only if the `textarea` has *more than one* line of text
treating single-line `textarea`s as regular `isSelectable` elements.