Remove code for tracking which frame is focused
Most of this isn't necessary anymore with the new implementation of nextFrame. UIComponents would
use this information to dismiss themselves when another frame was focused (see the removals in
ui_component.js), but I'm not sure this code was actually effectual in practice, and if so, we can do
without that functionality.