Rework how sendMessage message handlers work
When using sendMessage, each listener now enumerates which message types it wants to listen to.

We now consistently use `handler` to indicate the message type in the request object, rather than
`handler` and `name`. Previously, `name` was used to indicate that a content script should handle
the message, and `handler` to indicate that a background script should handle it. This was
undocumented and confusing.