Remove window unload handler
I believe this handler was previously used to help the background page track the state of a
tab/frame. We no longer maintain that state. Now, it was being used to unload content scripts when a
page navigates. However, I don't think this detection was reliable (unload is not called in all
cases), and so we now instead disable Vimium's content script whenever an event (keystroke or other)
that we're listening for occurs.

Adding a window unload handler can throw an exception when the page has unload disabled via
Permission-Policy: unload. See

https://groups.google.com/a/chromium.org/g/blink-dev/c/38Dpu-uhwFc