Completions: only use the custom search completer for custom searches
For simplicity.

This removes some functionality where if the user has a tab already open with the result of a
particular custom search, that tab won't be shown in the Vomnibar search results. I think that level
of UX polish is not worth the amount of code complexity it incurs.