Initialize the extension in the body of main.js
This resolves an issue where the extension would get into a bad state after disabling and then
re-enabling it in the Chrome extensions page. The keymapping for commands would never get written to
storage.session, and so no keybindings would work.

This issue did not appear to manifest in Firefox for some reason.