Improve how to detect whether to show an upgrade notification
Previously, we were storing the previous Vimium version in chrome.storage.sync.

1. This is not correct because that value should pertain to the local Vimium's state, but it's
   synced across all of the user's installations.
2. We don't need to keep this state. The browser provides this information in the onInstalled event.

Prior to this change, the upgrade from v1.67.x -> v2.0 was detected as a new install because the
schema of the settings data model has changed as part of v2.0, and so an upgrade notification was
not shown.