Use the tab.lastAccessed timestamp to sort tabs by recency in Vomnibar
This has been available in Firefox since Firefox 56, and was just added to Chrome 121.

As part of this, we no longer need to maintain state in the background service worker. After
migrating to manifest v3, this state was getting lost whenever the service worker gets unloaded.
As a result, this should fix #4368.