Options page: don't blur the current field after ctrl-enter to save options
This has been in from the beginning. I'm not sure why. One reason not to have this is that now that
we're going to show validation errors on the options page, one can save the options to check the
syntax of their key mapping config by typing ctrl-enter, see if there are any validation
messages, and then continue typing.