Link hints: fix an error when link hints in the help dialog on extension pages
For some reason this only surfaced an error in Firefox. On an extension page, frameId was set to
null, so all link hints were considered local for every frame, which led to errors.

frameId was both a local variable and later assigned to the globalScope, which was confusing anyway,
so this is an improvement in clarity.