setZoom command: require the level argument to be named
E.g. `map zz setZoom level=1.5`

This breaks backwards compatibility with existing configs, but hopefully not many people have this
in their configs yet as it was only recently introduced.

The reason for this change is to make it more manageable to document and validate our option
arguments. I think we should generally try to avoid unnamed/positional options because they
become confusing when mixed with named options.

Currently createTab is the only command that allows an unnamed option: it takes one or more
arbitrary URLs.