Add noop command
This is useful to "unmap" native browser mappings; for example I
sometimes accidentally press <c-h> or <c-b> instead of <c-g> to find the
next match, which opens the sidebar which I then need to hide with the
mouse. It's pretty annoying.

With this, I can do:

	map <c-h> noop
	map <c-b> noop

And it won't do anything.

Tested locally in Firefox 145 and Chromium 143.0.7499.40

Fixes #4213